Extract test setting to reduce duplicates
- Fix: #451
