+++ src/Frontend/Components/AppContainer/__tests__/AppContainer.test.tsx
@@ -6,25 +6,8 @@
 import { AppContainer } from '../AppContainer';
 import React from 'react';
 import { renderComponentWithStore } from '../../../test-helpers/render-component-with-store';
-import { IpcRenderer } from 'electron';
-
-let originalIpcRenderer: IpcRenderer;
 
 describe('The AppContainer', () => {
-  beforeAll(() => {
-    originalIpcRenderer = global.window.ipcRenderer;
-    global.window.ipcRenderer = {
-      on: jest.fn(),
-      removeListener: jest.fn(),
-      invoke: jest.fn(),
-    } as unknown as IpcRenderer;
-  });
-
-  afterAll(() => {
-    // Important to restore the original value.
-    global.window.ipcRenderer = originalIpcRenderer;
-  });
-
   test('renders AppContainer', () => {
     renderComponentWithStore(<AppContainer />);
   });
