+++ src/Frontend/Components/AttributionColumn/__tests__/AttributionColumn.test.tsx
@@ -4,7 +4,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { screen } from '@testing-library/react';
-import { IpcRenderer } from 'electron';
 import React from 'react';
 import {
   FollowUp,
@@ -46,27 +45,7 @@ import {
   insertValueIntoTextBox,
 } from '../../../test-helpers/attribution-column-test-helpers';
 
-let originalIpcRenderer: IpcRenderer;
-
 describe('The AttributionColumn', () => {
-  beforeAll(() => {
-    originalIpcRenderer = global.window.ipcRenderer;
-    global.window.ipcRenderer = {
-      on: jest.fn(),
-      removeListener: jest.fn(),
-      invoke: jest.fn(),
-    } as unknown as IpcRenderer;
-  });
-
-  afterAll(() => {
-    // Important to restore the original value.
-    global.window.ipcRenderer = originalIpcRenderer;
-  });
-
-  afterEach(() => {
-    jest.clearAllMocks();
-  });
-
   test('renders TextBoxes with right titles and content', () => {
     const testTemporaryPackageInfo: PackageInfo = {
       attributionConfidence: DiscreteConfidence.Low,
@@ -579,22 +558,6 @@ describe('The AttributionColumn', () => {
   });
 
   describe('The ResolveButton', () => {
-    beforeAll(() => {
-      originalIpcRenderer = global.window.ipcRenderer;
-      global.window.ipcRenderer = {
-        on: jest.fn(),
-        removeListener: jest.fn(),
-        invoke: jest.fn(),
-      } as unknown as IpcRenderer;
-    });
-
-    beforeEach(() => jest.clearAllMocks());
-
-    afterAll(() => {
-      // Important to restore the original value.
-      global.window.ipcRenderer = originalIpcRenderer;
-    });
-
     test('saves resolved external attributions', () => {
       const testTemporaryPackageInfo: PackageInfo = {};
       const expectedSaveFileArgs: SaveFileArgs = {
