+++ src/Frontend/Components/NotSavedPopup/__tests__/NotSavedPopup.test.tsx
@@ -4,7 +4,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { fireEvent, screen } from '@testing-library/react';
-import { IpcRenderer } from 'electron';
 import React from 'react';
 import { ButtonText, PopupType, View } from '../../../enums/enums';
 import {
@@ -44,24 +43,7 @@ function setupTestState(
   targetView && store.dispatch(setTargetView(targetView));
 }
 
-let originalIpcRenderer: IpcRenderer;
-
 describe('NotSavedPopup and do not change view', () => {
-  beforeAll(() => {
-    originalIpcRenderer = global.window.ipcRenderer;
-    global.window.ipcRenderer = {
-      on: jest.fn(),
-      invoke: jest.fn(),
-    } as unknown as IpcRenderer;
-  });
-
-  beforeEach(() => jest.clearAllMocks());
-
-  afterAll(() => {
-    // Important to restore the original value.
-    global.window.ipcRenderer = originalIpcRenderer;
-  });
-
   test('renders a NotSavedPopup', () => {
     const { store } = renderComponentWithStore(<NotSavedPopup />);
     setupTestState(store);
@@ -116,21 +98,6 @@ describe('NotSavedPopup and do not change view', () => {
 });
 
 describe('NotSavedPopup and change view', () => {
-  beforeAll(() => {
-    originalIpcRenderer = global.window.ipcRenderer;
-    global.window.ipcRenderer = {
-      on: jest.fn(),
-      invoke: jest.fn(),
-    } as unknown as IpcRenderer;
-  });
-
-  beforeEach(() => jest.clearAllMocks());
-
-  afterAll(() => {
-    // Important to restore the original value.
-    global.window.ipcRenderer = originalIpcRenderer;
-  });
-
   test('renders a NotSavedPopup', () => {
     const { store } = renderComponentWithStore(<NotSavedPopup />);
     setupTestState(store, View.Attribution);
