+++ src/Frontend/integration-tests/all-views-tests/__tests__/all-views.test.tsx
@@ -10,12 +10,10 @@ import {
   clickOnFilter,
   EMPTY_PARSED_FILE_CONTENT,
   goToView,
-  mockElectronIpcRendererOn,
+  mockElectronBackend,
   openDropDown,
-  TEST_TIMEOUT,
 } from '../../../test-helpers/general-test-helpers';
 import { screen } from '@testing-library/react';
-import { IpcChannel } from '../../../../shared/ipc-channels';
 import { renderComponentWithStore } from '../../../test-helpers/render-component-with-store';
 import {
   ButtonText,
@@ -23,40 +21,11 @@ import {
   FilterType,
   View,
 } from '../../../enums/enums';
-import { IpcRenderer } from 'electron';
 import { ParsedFileContent } from '../../../../shared/shared-types';
 import React from 'react';
 import { clickOnCardInAttributionList } from '../../../test-helpers/package-panel-helpers';
 
-let originalIpcRenderer: IpcRenderer;
-
-jest.setTimeout(TEST_TIMEOUT);
-
-function mockElectronBackend(mockChannelReturn: ParsedFileContent): void {
-  window.ipcRenderer.on
-    // @ts-ignore
-    .mockImplementation(
-      mockElectronIpcRendererOn(IpcChannel.FileLoaded, mockChannelReturn)
-    );
-}
-
 describe('The App integration', () => {
-  beforeAll(() => {
-    originalIpcRenderer = global.window.ipcRenderer;
-    global.window.ipcRenderer = {
-      on: jest.fn(),
-      removeListener: jest.fn(),
-      invoke: jest.fn(),
-    } as unknown as IpcRenderer;
-  });
-
-  beforeEach(() => jest.clearAllMocks());
-
-  afterAll(() => {
-    // Important to restore the original value.
-    global.window.ipcRenderer = originalIpcRenderer;
-  });
-
   test('app persists filters when changing views', () => {
     const mockChannelReturn: ParsedFileContent = {
       ...EMPTY_PARSED_FILE_CONTENT,
