+++ src/Frontend/integration-tests/audit-view-tests/__tests__/audit-view.test.tsx
@@ -9,12 +9,10 @@ import {
   EMPTY_PARSED_FILE_CONTENT,
   getOpenFileIcon,
   getParsedInputFileEnrichedWithTestData,
-  mockElectronIpcRendererOn,
-  TEST_TIMEOUT,
+  mockElectronBackend,
 } from '../../../test-helpers/general-test-helpers';
 import { App } from '../../../Components/App/App';
 import { screen } from '@testing-library/react';
-import { IpcRenderer } from 'electron';
 import React from 'react';
 import { IpcChannel } from '../../../../shared/ipc-channels';
 import {
@@ -63,35 +61,7 @@ import {
   expectReplaceAttributionPopupIsShown,
 } from '../../../test-helpers/popup-test-helpers';
 
-let originalIpcRenderer: IpcRenderer;
-
-jest.setTimeout(TEST_TIMEOUT);
-
-function mockElectronBackend(mockChannelReturn: ParsedFileContent): void {
-  window.ipcRenderer.on
-    // @ts-ignore
-    .mockImplementation(
-      mockElectronIpcRendererOn(IpcChannel.FileLoaded, mockChannelReturn)
-    );
-}
-
 describe('The App in Audit View', () => {
-  beforeAll(() => {
-    originalIpcRenderer = global.window.ipcRenderer;
-    global.window.ipcRenderer = {
-      on: jest.fn(),
-      removeListener: jest.fn(),
-      invoke: jest.fn(),
-    } as unknown as IpcRenderer;
-  });
-
-  beforeEach(() => jest.clearAllMocks());
-
-  afterAll(() => {
-    // Important to restore the original value.
-    global.window.ipcRenderer = originalIpcRenderer;
-  });
-
   test('renders TopBar and no ResourceBrowser when no resource file has been loaded', () => {
     renderComponentWithStore(<App />);
